Use RN modal to gain access to context; move delete functionality into ObsEditProvider
